pyproject.toml
[project]
name = "fast-healthchecks"
version = "0.0.0"
description = "FastHealthchecks"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    {name = "Vladislav Shepilov", email = "shepilov.v@protonmail.com"},
]
maintainers = [
    {name = "Vladislav Shepilov", email = "shepilov.v@protonmail.com"},
]
keywords = [
    "healthcheck", "library",
    "fastapi", "starlette", "faststream", "litestar",
    "asyncpg", "psycopg", "redis", "aio-pika", "httpx", "aiokafka", "motor",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
]
requires-python = "<4.0.0,>=3.10.0"
dependencies = []

[tool.setuptools]
packages = ["fast_healthchecks"]

[project.urls]
homepage = "https://github.com/shepilov-vladislav/fast-healthchecks"
documentation = "https://shepilov-vladislav.github.io/fast-healthchecks/"
source = "https://github.com/shepilov-vladislav/fast-healthchecks"
tracker = "https://github.com/shepilov-vladislav/fast-healthchecks/issues"

[project.optional-dependencies]
pydantic = ["pydantic>=2.10.3,<3.0.0"]
asyncpg = ["asyncpg>=0.30.0,<1.0.0"]
psycopg = ["psycopg>=3.2.3,<4.0.0"]
redis = ["redis>=5.2.1,<6.0.0"]
aio-pika = ["aio-pika>=9.5.3,<10.0.0"]
httpx = ["httpx>=0.28.1,<1.0.0"]
aiokafka = ["aiokafka>=0.12.0,<1.0.0"]
motor = ["motor>=3.6.0,<4.0.0"]
fastapi = ["fastapi[standard]>=0.115.6,<1.0.0"]
faststream = ["faststream>=0.5.33,<1.0.0"]
litestar = ["litestar>=2.13.0,<3.0.0"]
msgspec = ["msgspec>=0.18.6,<1.0.0"]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1,<5.0.0",
    "ruff>=0.8.2,<1.0.0",
    "mypy>=1.13.0,<2.0.0",
    "mypy-extensions>=1.0.0,<2.0.0",
    "pytest>=8.3.4,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-deadfixtures>=2.2.1,<3.0.0",
    "greenlet>=3.1.1,<4.0.0",
    "tox>=4.23.2,<5.0.0",
    "tox-uv>=1.16.1,<2.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "types-redis>=4.6.0.20241004,<5.0.0.0",
    "pytest-vcr>=1.0.2",
]
docs = [
    "mkdocs-include-markdown-plugin>=7.1.2",
    "mkdocs>=1.6.1,<2.0.0",
    "mkdocs-material>=9.5.47,<10.0.0",
    "mkdocstrings[python]>=0.27.0,<1.0.0",
    "pymdown-extensions>=10.12,<11.0",
]

[tool.uv]
default-groups = []

[tool.uv.sources]
msgspec = { git = "https://github.com/jcrist/msgspec.git", rev = "main" }

[tool.ruff]
line-length = 120
# Assume Python 3.10 to be compatible with typing.
target-version = "py310"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".ruff_cache",
    ".mypy_cache",
    ".tox",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "E",      # pycodestyle errors
    "C90",    # mccabe
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle, disabled because of we are using other docstring style
    "UP",     # pyupgrade
    "YTT",    # flake8-2020
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "BLE",    # flake8-blind-except
    "FBT",    # flake8-boolean-trap
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "CPY",    # flake8-copyright
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "DJ",     # flake8-django
    "EM",     # flake8-errmsg
    "EXE",    # flake8-executable
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "G",      # flake8-logging-format
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PYI",    # flake8-pyi
    "PT",     # flake8-pytest
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "INT",    # flake8-gettext
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "TD",     # flake8-todos
    "FIX",    # flake8-fixme
    "ERA",    # eradicate
    "PD",     # pandas-vet
    "PGH",    # pygrep-hooks
    "PL",     # Pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "NPY",    # NumPy-specific rules
    "AIR",    # Airflow
    "PERF",   # Perflint
    "FURB",   # refurb
    "LOG",    # flake8-logging
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by ruff-format
    "CPY001",  # Missing copyright notice at top of file
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-tidy-imports]
[tool.ruff.lint.flake8-tidy-imports.banned-api]
"pydantic_core.core_schema.FieldValidationInfo".msg = "FieldValidationInfo` is deprecated, use `ValidationInfo` instead."

[tool.ruff.lint.per-file-ignores]
"!fast_healthchecks/**.py" = ["D", "S", "SLF"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
minversion = "8.0"
addopts = "-ra -q"
filterwarnings = [
    "error",
    "error:::fast_healthchecks",
    "ignore:Failing to pass a value to the 'type_params' parameter of 'typing.ForwardRef._evaluate':DeprecationWarning",
]
markers = """
    integration: mark a test as an integration test
    unit: mark a test as a unit test
    imports: mark a test as an imports test
"""

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
disable_error_code = ["arg-type", "attr-defined"]
module = "tests.*"
ignore_missing_imports = true
check_untyped_defs = true

[tool.coverage.run]
source = ["fast_healthchecks"]

[tool.coverage.report]
precision = 1
include = ["fast_healthchecks/*"]
omit = [
    "tests/*",
]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",
    # Don't check obviously not implemented
    "raise NotImplementedError",
    # We don't really care what happens if fail
    "except ImportError:",
    # Don't check for typing-only code
    "if TYPE_CHECKING:",
    # Don't check for code that only runs itself
    "if __name__ == .__main__.:",
]
show_missing = true

[tool.tox]
legacy_tox_ini = """
    [tox]
    requires =
        tox>=4
    isolated_build = true
    envlist = py{310,311,312,313}-pydantic-{v1,v2}

    [testenv]
    deps =
        pydantic1: pydantic>=1.10.19,<2.0.0
        pydantic2: pydantic>=2.10.3,<3.0.0
    extras = ["test"]
    commands =
        pytest --dead-fixtures
        pytest --junitxml={envname}_report.xml --cov --cov-append -m 'not imports and not integration'
"""

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = false
version_files = [
    "pyproject.toml:version",
    "fast_healthchecks/__init__.py:__version__",
]
